Reduce sampling API clutter
