+++ python/ray/tune/__init__.py
@@ -7,10 +7,12 @@ from ray.tune.tune import run_experiments, run
 from ray.tune.experiment import Experiment
 from ray.tune.registry import register_env, register_trainable
 from ray.tune.trainable import Trainable
-from ray.tune.suggest import grid_search, function, sample_from
+from ray.tune.suggest import grid_search
+from ray.tune.sample import (function, sample_from, uniform, choice, randint,
+                             randn)
 
 __all__ = [
     "Trainable", "TuneError", "grid_search", "register_env",
     "register_trainable", "run", "run_experiments", "Experiment", "function",
-    "sample_from"
+    "sample_from", "uniform", "choice", "randint", "randn"
 ]
